+++ cogs/librarian/librarian.py
@@ -228,10 +228,18 @@ class Librarian(rubbercog.Rubbercog):
     async def macaddress(self, ctx, mac: str):
         """Get information about MAC address"""
         apikey = self.config.get("maclookup_token")
+
         if apikey == 0:
-            return await self.text.get("maclookup", "no_token")
+            return await self.output.error(
+                ctx,
+                self.text.get("maclookup", "no_token"),
+            )
+
         if "&" in mac or "?" in mac:
-            return await self.text.get("maclookup", "bad_mac", mention=ctx.author.mention)
+            return await self.output.error(
+                ctx,
+                self.text.get("maclookup", "bad_mac", mention=ctx.author.mention),
+            )
 
         url = f"https://api.maclookup.app/v2/macs/{mac}?format=json&apiKey={apikey}"
         res = await self.fetch_json(url)
@@ -266,6 +274,57 @@ class Librarian(rubbercog.Rubbercog):
         if res["blockStart"] != res["blockEnd"]:
             block += f"\n`{res['blockEnd']}`"
         embed.add_field(name=self.text.get("maclookup", "block"), value=f'`{res["blockStart"]}`')
+
+        await ctx.send(embed=embed)
+
+    @commands.cooldown(rate=2, per=20, type=commands.BucketType.user)
+    # The API has limit of 45 requests per minute
+    @commands.cooldown(rate=45, per=55, type=commands.BucketType.default)
+    @commands.command(aliases=["iplookup"])
+    async def ipaddress(self, ctx, query: str):
+        """Get information about an IP address or a domain name"""
+        if "&" in query or "?" in query:
+            return await self.output.error(
+                ctx,
+                self.text.get("iplookup", "bad_query", mention=ctx.author.mention),
+            )
+
+        url = (
+            f"http://ip-api.com/json/{query}"
+            "?fields=query,status,message,country,regionName,city,lat,lon,isp,org"
+        )
+        res = await self.fetch_json(url)
+        # TODO The API states that we should be listening for the `X-Rl` header.
+        # If it is `0`, we must stop for `X-ttl` seconds.
+        # https://ip-api.com/docs/api:json
+
+        if res["status"] == "fail":
+            embed = self.embed(
+                ctx=ctx,
+                title=self.text.get("iplookup", "error"),
+                description="`" + res["message"] + "`",
+                footer="ip-api.com",
+            )
+            return await ctx.send(embed=embed)
+
+        embed = self.embed(ctx=ctx, title=res["query"], footer="ip-api.com")
+        embed.add_field(
+            name=res["city"],
+            value=f"{res['regionName']}, {res['country']}",
+            inline=False,
+        )
+        embed.add_field(
+            name=self.text.get("iplookup", "geo"),
+            value=f"{res['lon']}, {res['lat']}",
+        )
+        embed.add_field(
+            name=self.text.get("iplookup", "org"),
+            value=res["org"],
+        )
+        embed.add_field(
+            name=self.text.get("iplookup", "isp"),
+            value=res["isp"],
+        )
 
         await ctx.send(embed=embed)